etf-trading-engine/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.11.0
statsmodels>=0.14.2
pyyaml>=6.0.1
requests>=2.32.0
